Move ResponsiveAppBar to root of components
