+++ frontend/src/components/landing/Landing.jsx
@@ -6,7 +6,7 @@ import "@fontsource/roboto/400.css";
 import "@fontsource/roboto/500.css";
 import "@fontsource/roboto/700.css";
 
-import ResponsiveAppBar from "./ResponsiveAppBar";
+import ResponsiveAppBar from "../ResponsiveAppBar";
 import {Card} from "@mui/material";
 
 export default function Landing() {
